Basics (#2)
* build: Now including GTFOProfiler, a custom profiler.

* chore: Updated formatting and ignored files

* build: Updated build files to reflect new changes

* feat: Changed to use Vulkan RAII API and other misc. additions.
Added GTFOProfiler, a custom homebrew instrumentation profiler.
Added basic spdlog logging functionality.

* fix: Profiling files were using .log ext. instead of .json

* feat: Error logs are now using stderr
Before errors and regular log messages were directed into stdout.

* chore: Removed anon. scope

* feat: Added GTFOProfiler profiling to createInstance()
This is very much unecessary for this stage of development but I wanted
to implement this just for fun.

* feat: Changed CXX_Standard to C++20

* refactor: Application now calls one init() rather then seperate ones.

* chore: Removed unused System class

* feat: Create Logic Device and Graphics Queue

* feat: Created a Swap Chain and Image Views
Completed the "Presentation" sub-section of the "Drawing a triangle"
section of the Khronos Vulkan Tutorial

* feat: Setup the graphics pipeline!

* feat: Hello, Triangle!